Refactor form submission modals
