+++ src/components/FormSubmissionModal/index.js
@@ -1,0 +1,2 @@
+export { default as FormSubmissionModal } from './FormSubmissionModal';
+export { default as SubmissionSuccessModal } from './SubmissionSuccessModal';
